requirements.txt
django==5.1.4
djangorestframework==3.15.2
Pillow==11.0.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
django-cors-headers==4.6.0
requests==2.32.3
python-dotenv==0.0.5
celery==5.4.0
redis==5.2.1